Add topic_prefix config to dynamically include composite stream topics in config
Streams are composed of multiple topics (in Kafka).  We want a way to easily map
from stream names to topics for easier automated consumption of active streams
in Kafka. Topic prefixes are often used to group topics together into a stream
(usually for datacenter replication).

Since StreamConfig's stream settings may be a regex, getting the topics list
using prefixes isn't always straightforward.  This patch adds a topics()
method to StreamConfig.  If a $stream name is passed in, then the topics
list returned will be prefixes with each configured topic_prefix.
Otherwise, the value of stream() will be used to construct the topic prefix.
If stream() (or a provided $stream) is a regex, then instead of a list of static topics,
topics() will return a list with a single regex with the topic prefixes in the regex.
E.g. /^mediawiki\.job\..+/ with prefixes eqiad. and codfw. becomes /^(eqiad\.|codfw\.)mediawiki\.job\..+/

A topics regex is still useful, as Kafka consumers often support subscribing to topics
via regexes.

Bug: T251609
Change-Id: I03fc69135549e1fd38e749fc7e34479e6fe4df8d