Recursively set array type 'assoc' on ApiStreamConfig results
Provide a hint to the API that the array containing the requested
stream configs (and all of its child properties) should be considered
associative arrays, and should therefore be represented in the
response as JSON objects rather than arrays, even if empty.

Exclude keys known to refer to list types.

Bug: T259917
Change-Id: I3f4cdf1580bfc45bd198898610f5053a7f5dd6f0