Expect wgEventStreams to be an associative array
Updates the wgEventStreams description and the StreamConfigs class
to require that wgEventStreams is an associative array. The primary
benefit of this change is that it will allow us to easily vary the
configuration for particular streams by wiki in WMF's MediaWiki
environments. Additionally, it improves performance by enabling a
constant-time lookup when a config is requested for a particular
stream that has a non-regex config.

For backward compatibility, the stream name (or regex) will be taken
from the key and will continue to be included as the 'stream' property
in the stream config itself. Integer indexed MW configs will
continue to work, as long as they have a 'stream' setting.

N.B. Surprisingly, PHP associative arrays *are ordered*[1][2], so
this change does not disturb the ordering behavior currently relied
upon for regex fallbacks.

[1] https://www.php.net/manual/en/language.types.array.php
[2] https://stackoverflow.com/questions/10914730/are-php-associative-arrays-ordered

Bug: T277193
Change-Id: I016146e7102cdc81641be92416dcbe42f47e92d5
Depends-On: I47ea9b7f2530e4574a3031ac75baec082f9e3d20
Depends-On: I2c87e7232be6a31ff9b6c002038f9c77459e825e