API: Only serialize streams as objects
The streamconfigs action API query module automatically serializes
anything object-like as an object. This behavior was introduced in
T259917 to fix the case where empty stream configs were being serialized
as lists, i.e.

{
  "streams": {}
}

{
  "streams": {
    "foo": {}
  }
}

However, the behavior is overly prescriptive. For example, the Metrics
Platform Client (MPC) producer configuration includes lists and objects,
which should be serialized as lists and maps respectively.

This patch limits the behavior described above to the top-level and
first-level parts of the response and anything beyond that, e.g.
streams[].producers is left to the MediaWiki Action API to serialize
appropriately.

Bug: T323032
Change-Id: I2de828679a4bd88f088a3697c5a97a94c6ec477b