API: Increase streams param values limit to 250
EventGate can be configured to refresh its internal stream configs cache
periodically. It does this by clearing the cache and then fetching
stream configs for all streams that were in it.

There are 123 streams with
destination_event_service=eventgate-analytics-external [0]. 122 of those
streams will have canary events submitted to them every 15 minutes [1].
The corresponding EventGate instance instance is configured to refresh
its internal stream configs cache every 5 minutes. So every 15 minutes,
it will clear its cache and then fetch the configs for all 122 streams
at once.

The default limit for the number of values for multi-value parameters is
50. The streams parameter has this same limit. If EventGate fetches
configs for 51 or more streams, then it will fail, leaving the cache
empty.

EventGate should be patched to fetch the stream configs in batches. For
now, increase the limit on the number of values for the streams
parameter.

[0] curl -s "https://meta.wikimedia.org/w/api.php?format=json&action=streamconfigs&constraints=destination_event_service=eventgate-analytics-external" | jq ".streams | length"
[1] curl -s "https://meta.wikimedia.org/w/api.php?format=json&action=streamconfigs&constraints=destination_event_service=eventgate-analytics-external" | jq ".streams | to_entries | map( select( .value.canary_events_enabled == true ) ) | length"

Bug: T326002
Change-Id: Ib8de24c12b70edf097b383f21df722e30019a762